chore: Refactor GetFolderTreeRequestHandler to include group teachers in query
The code changes in this commit modify the `GetFolderTreeRequestHandler` class to include group teachers in the query. This is achieved by introducing a new `GroupTeachersByUserIdSpec` class and updating the `GetFolderTreeRequestHandler` constructor to fetch group teachers based on the user ID. The retrieved group teacher IDs are then used in the `QuestionFoldersWithPermissionsSpecByUserId` query to filter question folders. This enhancement aims to improve the functionality and performance of the code.